refactor(main): Move initialization logic to fetcher package init
